refactor(public): move sidebar template under registry
Stage 3 of Phase 4. Lifts the existing PlanLandingView into
templates/sidebar-tree/SidebarLandingView so the registry's lazy import
resolves to real content instead of a Stage 2 placeholder.

No visual behavior change for existing districts — the file moved but
the rendered markup and routing are byte-for-byte identical. The Stage 2
placeholder that said "wiring lands in Phase 4 Stage 3" is now gone.

Also updates the sidebar group's page.tsx and the moved test to
reference the new name / default export shape expected by the registry.
