+++ src/app/district/[slug]/(sidebar)/page.tsx
@@ -1,11 +1,11 @@
 'use client'
 import dynamic from 'next/dynamic'
 
-const PlanLandingView = dynamic(
-  () => import('@/views/v2/public/PlanLandingView').then((m) => ({ default: m.PlanLandingView })),
+const SidebarLandingView = dynamic(
+  () => import('@/views/v2/public/templates/sidebar-tree/SidebarLandingView'),
   { ssr: false }
 )
 
 export default function PlanLandingPage() {
-  return <PlanLandingView />
+  return <SidebarLandingView />
 }
